Fix transaction handling. Add tests and demo app.
By default, peavy should be doing its database logging on a different
connection than the application uses, so its commits don't interfere
with the application.

Added unit tests. Well, test, singular, to verify that the logging is
persistent even if the app transaction is rolled back, and that logging
doesn't commit app writes, thus preventing them from being rolled back.

Added a demo app to show how to configure and use it, and to show the
UI. Maybe it'll light a fire under me to improve the UI.