Give up on using a single database.
Django 1.5 is making it impossible to test, as far as I can tell; you
get a circular dependency error that I wasn't able to work around.

It's not the best strategy anyway; it's best to put logging in a
separate database, perhaps even on different storage or a different
server. You'll regret a single database when something goes haywire and
you want to clean up the resultant logging and reclaim the space,
without affecting the rest of the site.